week-5/day-5/bike_store/requirements.txt
asgiref==3.6.0
Django==4.2
Faker==18.5.1
psycopg2-binary==2.9.6
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.4
